+++ src/app.js
@@ -1,176 +1,171 @@
 (function(jQuery, mbrApp) {
 
-	var curr;
+	var curr = null;
+	var editorHTML, editorCSS;
     mbrApp.regExtension({
-        name: "code-editor",
-        events: {		
-            load: function() {
-                var a = this;
-				a.$body.find(".navbar-devices").append('<li class="updatebutton" style="width:66px; height:58px; cursor:pointer" data-tooltipster="bottom" title="Reload Mobirise"><i class="mbr-icon-update mbr-iconfont"></i></li><li class="editbutton" style="width:66px; height:58px; cursor:pointer" id="editbtn" data-tooltipster="bottom" title="Code Editor"><i class="mbr-icon-code mbr-iconfont"></i></li>');
-				a.$body.on("click", ".updatebutton", function(b) {
-					Bridge.reload();
+        name: "witsec-code-editor",
+        events: {
+            beforeAppLoad: function() {
+                mbrApp.Core.addFilter("prepareComponent", function(a, b) {
+					// 'a' is the component window's HTML as string. We need to jQuery that, so we can do magic stuff with it
+					var h = jQuery(a);
+
+					// Add edit button to component buttons
+					var btn = '<span class="mbr-btn mbr-btn-default mbr-icon-code witsec-code-editor-editbutton" data-tooltipster="bottom" title="Edit Code"></span><style>.witsec-code-editor-editbutton:hover { background-color: #42a5f5 !important; }</style>';
+					if (h.find(".component-params").length)
+						h.find(".component-params").before(btn);
+					else if (h.find(".component-remove").length)
+						h.find(".component-remove").before(btn);
+
+					// Get the HTML as a string, then return that
+					a = h.prop("outerHTML");
+					h.remove();
+					return a;
 				});
-				a.$body.on("click", ".editbutton", function(b) {
-					var currentPage = mbrApp.Core.currentPage;		
-					var complist = '';
-					var curr = mbrApp.Core.resultJSON[currentPage].components[0];
-					for (x in mbrApp.Core.resultJSON[currentPage].components){
-						compname = mbrApp.Core.resultJSON[currentPage].components[x]._name
-						complist = complist + '<option value="'+ x +': '+compname +'">'+ x +': '+ compname + '</option>';
-					}
-					mbrApp.$body.on("change",'[name="blockSelect"]',function(a) {
-						var ind = this.selectedIndex;
-						var currentPage = mbrApp.Core.currentPage;
-						curr = mbrApp.Core.resultJSON[currentPage].components[ind];
-						jQuery("#htmlcode").text(curr._customHTML);
-						hljs.highlightBlock(document.getElementById("htmlcode"));
-						jQuery(".styles").val(JSON.stringify(curr._styles, undefined, 2));
-						$("#blockCid").val(curr._cid);
-						$("#blockAnchor").val(curr._anchor);
-					});
-					mbrApp.showDialog({
-						title: "Code Editor",
-						className: "dp-modal",
-						body: ['',
-						'<div class="row">',
-						'  <div class="col-lg-4"><select class"form-control blockselect" name="blockSelect">',complist,'</select></div>',
-						'  <div class="col-lg-3">_cid: <input type="text" class"form-control blockcid" id="blockCid" value="',curr._cid,'"></div>',
-						'  <div class="col-lg-5">_anchor: <input type="text" class"form-control blockanchor" id="blockAnchor" value="',curr._anchor,'"></div>',
-						'</div>',
-						'<div class="row">&nbsp;</div>',
-						'<div class="row">',
-						'  <div class="col-lg-12">',
-						'    <ul class="nav nav-tabs">',
-						'      <li class="active"><a data-toggle="tab" href="#html">HTML</a></li>',
-						'      <li><a data-toggle="tab" href="#style">STYLE</a></li>',
-						'    </ul>',
-						'    <div class="tab-content">',
-						'      <div id="html" class="tab-pane fade in active">',
-						'        <pre><code class="html" id="htmlcode" contenteditable="true">',escHTML(curr._customHTML),'</code></pre>',
-						'        <button type="button" class="btn btn-primary" id="searchhtml">SEARCH-REPLACE</button>',
+			},
+
+            load: function() {
+				var a = this;
+
+				// Do stuff to load all required js files (doing it this way somehow ensures all files are loaded properly, while using params.json often messes things up...)
+				var dir = mbrApp.getAddonDir("witsec-code-editor");
+				var scrarr = ["codemirror-5.44.0.min.js", "jshint-2.10.2.min.js", "lint-javascript-lint-5.44.0.min.js", "lint-lint-5.44.0.min.js", "addon-active-line-5.44.0.min.js", "mode-css-5.44.0.min.js", "mode-htmlmixed-5.44.0.min.js", "mode-javascript-5.44.0.min.js", "mode-xml-5.44.0.min.js"];
+
+				var scrhtml = "";
+				scrarr.forEach(s => {
+					scrhtml += '<script src="' + dir + '/lib/' + s + '"></script>\n';
+				});
+
+				// Create the skeleton for the overlay and edit fields
+				if (!$("#witsec-code-editor").length) {
+					a.$body.append([
+						'<div id="witsec-code-editor">',
+						scrhtml,
+						'  <div>',
+						'    <div class="row witsec-code-editor-row">',
+						'      <div class="col-lg-8 witsec-code-editor-col">',
+						'        <div class="witsec-code-editor-header"><h4>HTML</h4></div>',
+						'        <textarea id="witsec-code-editor-html"></textarea>',
 						'      </div>',
-						'      <div id="style" class="tab-pane fade">',
-						'        <textarea class="form-control styles" id="styles" rows="23" style="font-family:Consolas; background-color:#333; color:#99ec99">',
-						         JSON.stringify(curr._styles, undefined, 2),
-						'        </textarea>',
-						'        <button type="button" class="btn btn-primary" id="searchstyles">search-replace</button>',
-						'        <button type="button" class="btn btn-primary" id="validjson">VALIDATE JSON</button>',
-						'        <button type="button" class="btn btn-primary" id="tojson">CONVERT CSS TO JSON</button>',
-						'      </div>',
+						'      <div class="col-lg-4 witsec-code-editor-col">',
+						'        <div class="witsec-code-editor-header"><h4>CSS (JSON format)</h4></div>',
+						'        <textarea id="witsec-code-editor-css"></textarea>',
+						'      <div>',
 						'    </div>',
 						'  </div>',
-						'</div>',
-						'<script>hljs.highlightBlock(document.getElementById("htmlcode"));</script>',
-						'<style>pre code {height: 450px;} pre.code {overflow: auto;}</style>'
-						].join("\n"),
-						buttons: [{
-							label: "SAVE",
-							default: !0,
-							callback: function() {
-								try {
-									curr._customHTML = jQuery("#htmlcode").text();
-									curr._styles = JSON.parse(jQuery(".styles").val());
-									mbrApp.runSaveProject(function() {
-										mbrApp.loadRecentProject(function(){
-											$("a[data-page='"+currentPage+"']").trigger("click")
-										});
-									});
-								}
-								catch(err){
-									mbrApp.alertDlg(err.name + ' with message : ' +err.message);
-								}
-							}
-						},
-							{
-							label: "CLOSE",
-							default: !0,
-							callback: function() {									
-								jQuery(".dp-modal").remove();
-								mbrApp.$body.off("click", ".dp-modal")
-							}
+						'  <button class="witsec-code-editor-save btn btn-fab btn-raised btn-primary" data-tooltipster="top" title="Save"><i class="mbr-icon-success"></i></button>',
+						'  <button class="witsec-code-editor-cancel btn btn-fab btn-raised btn-material-red" data-tooltipster="top" title="Cancel"><i class="mbr-icon-close"></i></button>',
+						'<div>'
+					].join("\n"));
+
+					// Initialize CodeMirror for the HTML Editor
+					editorHTML = CodeMirror.fromTextArea(a.$body.find("#witsec-code-editor-html")[0], {
+						styleActiveLine: true,
+						lineNumbers: true,
+						lineWrapping: false,
+						mode: "htmlmixed",
+						theme: "solarized dark",
+						lint: true,
+						gutters: ["CodeMirror-lint-markers"]
+					});
+
+					// Initialize CodeMirror for the CSS Editor
+					editorCSS = CodeMirror.fromTextArea(a.$body.find("#witsec-code-editor-css")[0], {
+						styleActiveLine: true,
+						lineNumbers: true,
+						lineWrapping: false,
+						mode: "application/json",
+						theme: "solarized dark",
+						lint: true,
+						gutters: ["CodeMirror-lint-markers"]
+					});
+				}
+
+				// Click handler for "edit code" icon
+				a.$template.on("click", ".witsec-code-editor-editbutton", function(e) {
+					// Find the index of the clicked icon
+					a.$template.find('.witsec-code-editor-editbutton').each(function(index, obj) {
+						if (e.target == obj) {
+							curr = mbrApp.Core.resultJSON[mbrApp.Core.currentPage].components[index];
 						}
-						]
 					});
-					$("#searchhtml").on("click", function(b) {
-						b.preventDefault();
-						var selectText = window.getSelection().toString();
-						mbrApp.showDialog({
-							title: "Search & Replace",
-							className: "replace-modal",
-							body: ['<form class="page-settings-form">\n<div class="form-group clearfix">\n<div class="col-md-12">\nsearch for :<input class="form-control inputmodal" type="text" id="searchstr" placeholder="search for" value="search for"/>\nreplace by:<input class="form-control inputmodal" type="text" id="replacestr" placeholder="replace by" value=""/>\n</div>\n</div>\n</form>'].join("\n"),
-							show: function(g) {
-							$("#searchstr").val(selectText);
-							},
-							buttons: [{
-								label: "Replace all",
-								default: !0,
-								callback: function() {
-										var replaceText = $("#replacestr").val();
-										selectText = $("#searchstr").val();
-										var newText = $('#htmlcode').text().split(selectText).join(replaceText);
-										$('#htmlcode').text(newText);
-										hljs.highlightBlock(document.getElementById("htmlcode"));
-								}
-							},
-							{
-								label: "Close/Cancel",
-								default: !0,
-								callback: function() {
-									$(".replace-modal").remove()
-								}
-							}
-							]
-						})
-					});
-					$("#searchstyles").on("click", function(b) {
-						b.preventDefault();
-						var selectText = document.getElementById("styles").value.substring(document.getElementById("styles").selectionStart,document.getElementById("styles").selectionEnd);
-						mbrApp.showDialog({
-							title: "Search & Replace",
-							className: "replace-modal",
-							body: ['<form class="page-settings-form">\n<div class="form-group clearfix">\n<div class="col-md-12">\nsearch for :<input class="form-control inputmodal" type="text" id="searchstr" placeholder="search for" value="search for"/>\nreplace by:<input class="form-control inputmodal" type="text" id="replacestr" placeholder="replace by" value=""/>\n</div>\n</div>\n</form>'].join("\n"),
-							show: function(g) {
-							$("#searchstr").val(selectText);
-							},
-							buttons: [{
-								label: "Replace all",
-								default: !0,
-								callback: function() {
-										var replaceText = $("#replacestr").val();
-										selectText = $("#searchstr").val();
-										var newText = $('#styles').val().split(selectText).join(replaceText);
-										//alert(newText);
-										$('#styles').val(newText);
-								}
-							},
-							{
-								label: "Close/Cancel",
-								default: !0,
-								callback: function() {
-									$(".replace-modal").remove()
-								}
-							}
-							]
-						})
-					});
-					$('#blockCid').on("change", function () {
-						curr._cid = $('#blockCid').val();
-					});
-					$('#blockAnchor').on("change", function () {
-						curr._anchor = $('#blockAnchor').val();
-					});
-					$('#validjson').on("click", function () {
-						validatejson(jQuery(".styles").val());
-					});
-					$('#tojson').on("click", function () {
-						css = jQuery(".styles").val();
-						jsoncss = tojson(css);
-						validatejson(jsoncss);
-						curr._styles = JSON.parse(jsoncss);							
-						jQuery(".dp-modal").remove();
-						$(".blockedit").click();							
-					});						
-				});				
+
+					// If curr is null, something is wrong
+					if (curr === null) {
+						mbrApp.alertDlg("An error occured while opening the Code Editor.");
+						return false;
+					}
+
+					// If no _customHTML exists, it's probably a "component.js" kinda block, which doesn't have editable HTML
+					if (!curr._customHTML) {
+						mbrApp.alertDlg("Sorry, this block can't be edited with the Code Editor.");
+						return false;
+					}
+
+					// Fill HTML and CSS here...
+					editorHTML.setValue(curr._customHTML);
+					editorCSS.setValue(JSON.stringify(curr._styles, undefined, 2));
+
+					// Clear history, so ctrl-z doesn't empty the editors
+					editorHTML.clearHistory();
+					editorCSS.clearHistory();
+
+					// Set editor columns to max height available
+					a.$body.find(".witsec-code-editor-col").height(window.innerHeight - 40);	// 40 is the height of the header
+
+					// Set editors to fill 100% of the div they're in
+					editorHTML.setSize("100%", "100%");
+					editorCSS.setSize("100%", "100%");
+
+					// Make the editor appear
+					$("#witsec-code-editor").height("100%");
+				});
+
+				// Auto-resize the editor columns if needed
+				$(window).resize(function() {
+					// If the Code Editor is visible, set editor columns to max height available
+					if (a.$body.find("#witsec-code-editor").height() != "0") {
+						a.$body.find(".witsec-code-editor-col").height(window.innerHeight - 40);	// 40 is the height of the header
+					}
+				});
+
+				// Save everything
+				a.$body.on("click", ".witsec-code-editor-save", function(b) {
+					try {
+						// Check if the CSS (JSON) contains any errors
+						try {
+							var styles = JSON.parse(editorCSS.getValue());
+						}
+						catch(e) {
+							mbrApp.alertDlg("The CSS doesn't appear to be in proper JSON format.");
+							return false;
+						}
+
+						// Grab the HTML and save both HTML and CSS to curr
+						curr._customHTML = editorHTML.getValue();
+						curr._styles = styles;
+
+						// Save
+						mbrApp.runSaveProject(function() {
+							mbrApp.loadRecentProject(function(){
+								$("a[data-page='" + mbrApp.Core.currentPage + "']").trigger("click")
+							});
+						});
+					}
+					catch(err){
+						mbrApp.alertDlg(err.name + ': ' + err.message);
+					}
+
+					// Make the editor disappear
+					$("#witsec-code-editor").height("0");
+					curr = null;
+				});
+
+				// Do nothing and hide the editor
+				a.$body.on("click", ".witsec-code-editor-cancel", function(b) {
+					$("#witsec-code-editor").height("0");
+					curr = null;
+				});
             }
         }
     })