+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
-## 0.1 (2019-03-25)
+## 0.2 (2019-03-31)
+
+- Added side-by-side editors, one for HTML, one for CSS
+- Updated the way the Javascript libraries are loaded. This solves the bug where code highlighting didn't always work, or active line didn't show, etc.
+- Added check if a block's "_customHTML" is available. If it isn't, the user receives a warning that the editor can't be used to edit said block
+- Added "edit" icon to each block
+- Removed "editor" icon from Mobirise navbar
+- Removed bunch of (now) unused coding
+
+## 0.1 (unreleased)
 
 This is the initial release of the Code Editor extension for Mobirise, forked from DeltaPi's Code Editor.